+++ cypress/e2e/02-signin.cy.ts
@@ -1,0 +1,11 @@
+import { createRegistryConfig, signinTests } from '@verdaccio/e2e-ui';
+
+const registryUrl = Cypress.env('VERDACCIO_URL') || 'http://localhost:4873';
+
+const config = createRegistryConfig({
+  registryUrl,
+  title: 'Verdaccio e2e',
+  credentials: { user: 'test', password: 'test' },
+});
+
+signinTests(config);
