+++ cypress/e2e/06-publish.cy.ts
@@ -1,0 +1,11 @@
+import { createRegistryConfig, publishTests } from '@verdaccio/e2e-ui';
+
+const registryUrl = Cypress.env('VERDACCIO_URL') || 'http://localhost:4873';
+
+const config = createRegistryConfig({
+  registryUrl,
+  title: 'Verdaccio e2e',
+  credentials: { user: 'test', password: 'test' },
+});
+
+publishTests(config);
