--- e2e/ui/config/config.yaml
@@ -1,41 +1,0 @@
-# Verdaccio config used by the e2e-ui CI job.
-# Tuned for the suites exported by @verdaccio/e2e-ui — keep these flags in sync
-# with the requirements documented in the e2e-ui README.
-
-storage: ./storage
-plugins: ./plugins
-
-web:
-  enable: true
-  title: Verdaccio
-  login: true
-  showSettings: true
-
-auth:
-  htpasswd:
-    file: ./htpasswd
-
-# Bumped from the default 1000 / 15min so a full Cypress run does not 429
-# against /-/verdaccio/sec/login or /-/user/*. Verdaccio reads this at
-# startup only — restart after editing.
-userRateLimit:
-  windowMs: 1000
-  max: 10000
-
-uplinks:
-  npmjs:
-    url: https://registry.npmjs.org/
-
-packages:
-  '@*/*':
-    access: $all
-    publish: $anonymous $authenticated
-    unpublish: $anonymous $authenticated
-    proxy: npmjs
-  '**':
-    access: $all
-    publish: $anonymous $authenticated
-    unpublish: $anonymous $authenticated
-    proxy: npmjs
-
-log: { type: stdout, format: pretty, level: info }
