--- e2e/ui/package.json
@@ -1,17 +1,0 @@
-{
-  "private": true,
-  "name": "verdaccio-e2e-ui-tests",
-  "version": "0.0.0",
-  "description": "Cypress UI E2E tests for Verdaccio (powered by @verdaccio/e2e-ui)",
-  "devDependencies": {
-    "@verdaccio/e2e-ui": "2.3.0",
-    "cypress": "15.10.0"
-  },
-  "scripts": {
-    "cypress:install": "cypress install",
-    "cypress:open": "cypress open",
-    "test": "cypress install && cypress run",
-    "test:local": "node ./scripts/run-local.mjs",
-    "test:local:open": "node ./scripts/run-local.mjs --open"
-  }
-}
