+++ package.json
@@ -54,6 +54,7 @@
     "@types/supertest": "7.2.0",
     "@types/validator": "13.15.10",
     "@verdaccio/e2e-cli": "2.3.0",
+    "@verdaccio/e2e-ui": "2.3.0",
     "@verdaccio/eslint-config": "13.0.0",
     "@verdaccio/package-filter": "workspace:*",
     "@verdaccio/types": "workspace:*",
@@ -63,6 +64,7 @@
     "aria-query": "5.3.2",
     "concurrently": "9.2.1",
     "cross-env": "10.1.0",
+    "cypress": "15.10.0",
     "debug": "4.4.3",
     "eslint": "10.0.3",
     "eslint-plugin-cypress": "6.1.0",
@@ -107,9 +109,11 @@
     "lint": "eslint --max-warnings 90 .",
     "lint:fix": "eslint --fix .",
     "test": "pnpm --filter \"./packages/**\" test",
-    "test:e2e:ui": "pnpm --filter verdaccio-e2e-ui-tests test",
-    "test:e2e:ui:local": "pnpm --filter verdaccio-e2e-ui-tests test:local",
-    "test:e2e:ui:local:open": "pnpm --filter verdaccio-e2e-ui-tests test:local:open",
+    "e2e:ui:start": "node packages/verdaccio/bin/verdaccio --config ./scripts/e2e-ui-config.yaml",
+    "e2e:ui:run": "cypress run",
+    "e2e:ui:open": "cypress open",
+    "e2e:ui:local": "./scripts/e2e-ui-local.sh",
+    "e2e:ui:local:open": "./scripts/e2e-ui-local.sh --open",
     "start": "concurrently --kill-others \"pnpm _start:server\" \"pnpm _start:web\"",
     "contributors": "ts-node ./scripts/contributors-update.ts",
     "script:addson": "ts-node ./scripts/addon-update.ts",
