+++ pnpm-workspace.yaml
@@ -13,4 +13,3 @@ packages:
   - packages/plugins/local-storage
   - packages/plugins/package-filter
   - e2e/cli/*
-  - e2e/ui
